backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langgraph==0.0.60
langchain-groq==0.1.4
groq==0.8.0
pydantic==2.7.2
supabase==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
pypdf==4.2.0
python-multipart==0.0.9
